requirements.txt
slack_sdk
requests
msal
python-dotenv
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyGithub
gitpython